fix requirements + setup

+++ setup.py
@@ -11,18 +11,19 @@ def setup(mode = 'normal'):
 		print('FAIL. Please save your telebot token in TOKEN file')
 		return
 
-	if mode != 'restart':
+	if mode != 'restart' and mode != 'debug':
 		os.system('curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py')
 		os.system('python3 get-pip.py')
 		os.system('rm get-pip.py')
 
 		os.system('pip3 install -r requirements.txt')
 		os.system('pip3 install python-telegram-bot --upgrade') # need to use some experiement feature, e.g. message filtering
+		os.system('pip3 install export_to_telegraph --upgrade')
 
 	# kill the old running bot if any. If you need two same bot running in one machine, use mannual command instead
 	os.system("ps aux | grep python | grep export.py | awk '{print $2}' | xargs kill -9")
 
-	if arg == 'debug':
+	if mode == 'debug':
 		os.system(RUN_COMMAND[6:-2])
 	else:
 		os.system(RUN_COMMAND)
